+++ Client/src/components/game/GameUI.js
@@ -6,7 +6,6 @@
  */
 
 import axios from 'axios';
-import Xarrow from "react-xarrows";
 import Province from './Province';
 import React, { useState } from 'react';  
 import './GameUI.css';
@@ -15,8 +14,8 @@ import {
     from '../../functionality/pendingActions';
 
 export default function GameUI( 
-  {onSelectAction, onMergeArmies, pushPendingData,
-    names, owners, flavors, terrains, provinceId, armies, session, player, battle, pending} ) {
+  {onSelectAction, onMergeArmies, pushPendingData, getTime,
+    names, owners, flavors, terrains, provinceId, armies, session, player, battle} ) {
   
   const worldRowSize = Math.sqrt(session.world_size);
   const [mergeConfirmation, setMergeConfirmation] = useState(false);
@@ -101,7 +100,8 @@ export default function GameUI(
           || armies[2][toProvince] == null 
           || armies[3][toProvince] == null) {      
           console.log("move army " + army + " from province " + fromProvince + " to " + toProvince);
-          postMovement(province1, province2, fromProvince, toProvince, session, player, army, pushPendingData);
+          postMovement(province1, province2, fromProvince, toProvince, 
+            session, player, army, getTime, pushPendingData);
 
         } else {
           console.log("No available army slots in that province!");
@@ -109,7 +109,8 @@ export default function GameUI(
       // If the province is not ours, attack!
       } else {
         console.log("attack with army " + army + " from province " + fromProvince + " to " + toProvince);
-        postMovement(province1, province2, fromProvince, toProvince, session, player, army, pushPendingData);
+        postMovement(province1, province2, fromProvince, toProvince, 
+          session, player, army, getTime, pushPendingData);
       }
     } else {
       console.log("Province is too far away!");
@@ -189,38 +190,6 @@ export default function GameUI(
     <div className="main_screen">
         <BuildBody />
     </div>
-    <ShowMovement pending={pending} armies={armies} />
     </>
   );
-}
-
-
-function ShowMovement({pending, armies}) {
-  const arrows = [];
-  for (let i = 0; i < pending.length; i++) {
-    if (pending[i].type == 'movement') {
-      // Check which slot the army comes from
-      let slot = 1;
-      for (let j = 0; j < armies.length; j++) {
-        if (armies[j][pending[i].provinceN] == pending[i].army_id) {
-          slot = j+1;
-        }
-      }
-
-      const days = pending[i].end / 24 | 0;
-      const hours = pending[i].end % 24;
-
-      const startDestination = "province" + pending[i].provinceN + "_army" + slot;
-      const endDestination = "province" + pending[i].province2N + "_army1";
-      const endLabel = "Arrives at: " + days + ":" + hours; 
-
-      arrows.push(<Xarrow start={startDestination} end={endDestination} labels={{middle: endLabel}} />);
-    }
-  }
-  
-  return (
-    <>
-      {arrows}
-    </>
-  )
 }
